Remove Bit From Project

+++ DNSLab/Shared/Components/Page/PageCommentItem.razor.cs
@@ -1,8 +1,12 @@
 ﻿using DNSLab.DTOs.Comments;
+using DNSLab.Repository;
+using MudBlazor;
 
 namespace DNSLab.Shared.Components.Page;
 partial class PageCommentItem
 {
+    [Inject] private IDialogService DialogService { get; set; }
+
     [Parameter] public Guid AuthUserId { get; set; } = Guid.Empty;
     [Parameter] public PageCommentDTO PageComment { get; set; }
     [Parameter] public EventCallback ReplyButtonClick { get; set; }
@@ -11,26 +15,27 @@ partial class PageCommentItem
     bool IsEditing = false;
     bool IsDeleted = false;
 
-    Modal DeleteModal;
+    async Task DeleteComment()
+    {
+        bool? result = await DialogService.ShowMessageBox(
+            "هشدار",
+            $"آیا از حذف نظر ({PageComment.Text}) مطمئن هستید؟",
+            yesText: "حذف", cancelText: "انصراف");
 
-    CreatePageCommentDTO ReplyPageComment = new CreatePageCommentDTO { Text = String.Empty };
-
-    async Task DeleteComment() => await DeleteModal.Open();
+        if (result == true)
+        {
+            if (await _CommentRepository.RemovePageComment(PageComment.Id))
+            {
+                IsDeleted = true;
+            }
+        }
+    }
 
     async Task SubmitEditComment()
     {
         if (await _CommentRepository.UpdatePageComment(PageComment))
         {
             IsEditing = false;
-        }
-    }
-
-    async Task SubmitDeleteComment()
-    {
-        if (await _CommentRepository.RemovePageComment(PageComment.Id))
-        {
-            await DeleteModal.Close();
-            IsDeleted = true;
         }
     }
 }
